Ensure that IO::Socket::IP is DESTROY'd before can_bind returns
be316f1f worked around a core perl bug[1] which causes Perl < 5.24.0
to DESTROY the IO::Socket::IP object too late, causing the empty port
to not actually be empty if it was used immediately.

Unfortunately, this change was lost in ba745fd1, which again put the
return of `can_bind` on the stack along with the actual return value,
re-instating the bug.

Split the statement across two statements again, and add a comment on
the importance of preserving it.

[1] https://rt.perl.org/Public/Bug/Display.html?id=124248